+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/SerializationTest.kt
@@ -1,18 +1,9 @@
 package at.asitplus.wallet.lib.oidvci
 
-import at.asitplus.openid.AuthenticationRequestParameters
-import at.asitplus.openid.CredentialFormatEnum
-import at.asitplus.openid.CredentialRequestParameters
-import at.asitplus.openid.CredentialRequestProof
-import at.asitplus.openid.CredentialResponseParameters
-import at.asitplus.openid.OpenIdAuthorizationDetails
-import at.asitplus.openid.OpenIdConstants
+import at.asitplus.openid.*
 import at.asitplus.openid.OpenIdConstants.GRANT_TYPE_AUTHORIZATION_CODE
 import at.asitplus.openid.OpenIdConstants.GRANT_TYPE_CODE
 import at.asitplus.openid.OpenIdConstants.TOKEN_TYPE_BEARER
-import at.asitplus.openid.SupportedCredentialFormatDefinition
-import at.asitplus.openid.TokenRequestParameters
-import at.asitplus.openid.TokenResponseParameters
 import at.asitplus.wallet.lib.data.VcDataModelConstants.VERIFIABLE_CREDENTIAL
 import at.asitplus.wallet.lib.oidc.jsonSerializer
 import io.kotest.core.spec.style.FunSpec
@@ -20,6 +11,7 @@ import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
 import io.ktor.http.*
 import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.JsonPrimitive
 import kotlin.random.Random
 import kotlin.time.Duration.Companion.seconds
 
@@ -79,7 +71,8 @@ class SerializationTest : FunSpec({
 
     fun createCredentialResponse() = CredentialResponseParameters(
         format = CredentialFormatEnum.JWT_VC,
-        credential = randomString(),
+        credential = JsonPrimitive(randomString()),
+        credentials = listOf(JsonPrimitive(randomString())),
         transactionId = randomString(),
         clientNonce = randomString(),
         clientNonceExpiresIn = Random.nextInt(1, Int.MAX_VALUE).seconds,